packages/rlm-claude/pyproject.toml
[project]
name = "rlm-claude"
version = "0.1.0"
description = "RLM-powered codebase analysis for Beethoven"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "tiktoken>=0.5.0",
    "openai>=1.0.0",
]

[project.scripts]
rlm-claude = "rlm_claude.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rlm_claude"]

[tool.uv.sources]
# Local RLM package - not on PyPI
# Will be added to path manually in code
